requirements.txt
selenium
pytest
allure-pytest
pytest-html
python-dotenv